+++ lib/providers/orders.dart
@@ -4,7 +4,7 @@ import 'cart.dart';
 
 class OrderItem {
   final String id;
-  final double amount;
+  final int amount;
   final List<CartItem> products;
   final DateTime dateTime;
   bool status;
@@ -27,7 +27,7 @@ class Orders with ChangeNotifier {
     return _orders.length;
   }
 
-  void addOrder(List<CartItem> cartProducts, double total) {
+  void addOrder(List<CartItem> cartProducts, int total) {
     _orders.insert(
       0,
       OrderItem(
